Make possible to resend mail notifications for shares from the dropdown
When a file is shared with a user and that user has an e-mail set in her
profile a mail notification is sent to that e-mail. However, once a file
has been shared there is currently no way to send again that e-mail
notification if needed for any reason.

Users that are already sharees are no longer shown in the share
dropdown, so one option could be to provide an additional action in the
share menu; however, although this would work for a small list of
sharees it could be slow if there are a lot of them. Therefore, instead
of adding an action in the share menu, the share dropdown was modified
so if a user is already a sharee and she has an e-mail set in her
profile then the user will appear in the dropdown and clicking on it
will send again the e-mail notification.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>