Fix wrongly removing remote shares when an e-mail share was found
In the commit "Make possible to resend mail notifications for shares
from the dropdown" a change which wrongly removed a remote sharee from
the suggestions when an e-mail share was found. This commit fixes it
restoring the previous and right behaviour, that is, removing the e-mail
sharee from the suggestions.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>